Make effects run on elapsed time; add particles and shader tiers
Effects now behave the same whatever framerate the hardware renders at: the
same settings give the same motion per second on a 30 fps wall and a 5,000 fps
desktop, with the faster device simply drawing it more smoothly. Adds a
particle kernel and a shader tier to the power-function library, with five new
effects built on them.

desktop 140us/7,142fps | esp32 4,164us/240fps | flash: desktop 1,075KB (+1KB),
esp32 1,704KB, esp32s3-n16r8 1,699KB (+3KB)

Core
- math16: isqrt64 no longer overflows its first Newton step, which returned 0
  for UINT64_MAX and would have read as zero separation once collide() started
  using it. Seeded from a power-of-two bound instead.
- noise: 16-bit tier (inoise16 1/2/3D, fbm16). lerp16 widens its product to
  int64 (the 32-bit form was signed overflow on roughly a quarter of samples),
  and fbm16 sums octaves at full width rather than shifting each down by 8,
  which had made the output 8-bit wearing a 16-bit type: 195 distinct values
  over 20,000 samples, now 15,118.

Light domain
- particles.h: the kernel. SoA pool over caller-owned buffers, forces,
  semi-implicit Euler, walls, collisions, emitters, and FrameTime as the shared
  elapsed-time scale.
- shader.h: the GLSL vocabulary in fixed point, plus project/depthFade.
- raymarch.h: sphere tracing, gated on SOC_CPU_HAS_FPU so small fixtures get it
  too, not desktop only.
- collide() computes distance in int64 end-to-end, including the root, so a
  large contact radius no longer truncates.
- wrap() reduces by modulo instead of per-span loops: constant time whatever the
  overshoot, and behaviour-identical to the loops it replaces, including the
  asymmetric endpoints (a multiple lands on span from above, on 0 from below).
- Echo, Lissajous, Ballpit, Particles and StarField moved onto elapsed time.
  Echo spends every whole reference frame a tick covers rather than one, so
  below 60 fps the trail compounds at 60 Hz instead of at the render rate; it
  was measured running 1 feedback pass per second at 240 fps instead of 60.
  Particles carries its sub-step remainder, so a slow speed still moves at high
  fps rather than truncating to a standstill.
- New effects: VectorBalls, plus the particle and shader showcases.

Tests
- unit_Effects_framerate.cpp audits all 51 effects at 60 vs 1200 fps, with
  targeted cases for the two mechanisms the aggregate cannot see.
- Coverage for isqrt64 across the full 64-bit range, the 16-bit noise
  properties, wrap endpoints, scroll positions, fract on negatives, and the
  attractor's magnitude bound.
- The empty-grid live-modifier test now uses a modifier that returns true from
  hasModifyLive(), so it exercises the gate it names.
- Echo, BouncingBalls, Lissajous and Tetrix goldens recaptured: what they draw
  on a given frame changed, the motion per second did not.

Docs/CI
- power-functions.md gains the particles and shader sections; caller lists
  record Particles.
- architecture.md states the tick-rate rule and drops a duplicated paragraph.
- The migration rule lives with the library rather than on the effect catalog.
- Backlog: worley, and the remaining 16-bit noise tuning.

Reviews
- CodeRabbit, 15 findings: fixed the substantive ones; skipped the two grid
  guards, since Layer::tick already gates on hasGrid and a guard in an effect is
  both dead code and an orchestration violation.
- Reviewer (Fable), 10 findings + 4 nits: 8 fixed. Skipped map32's overflow, as
  it needs both spans near 2^32 and every call site maps extents or bytes;
  documented the bound instead. Skipped moving FrameTime to core and abstracting
  raymarch behind a platform interface: both are real, both are refactors across
  every caller rather than review fixes.

Known, unfixed: Fireworks still counts frames for its launch roll. It has two
framerate bugs that currently cancel (frame-counted launches, and a
fadeToBlackBy floor that erases the trail faster on a fast device); fixing one
exposes the other, and fixing both changes how the effect looks, so it gets its
own commit.

+++ src/light/effects/RingsEffect.h
@@ -86,7 +86,7 @@ public:
                     // the ring never appeared out there at all.
                     const uint32_t d = dist16(dx, dy);
                     int32_t diff = static_cast<int32_t>(d) - static_cast<int32_t>(radius_[i]);
-                    if (diff < 0) diff = static_cast<int16_t>(-diff);
+                    if (diff < 0) diff = -diff;   // stays int32: narrowing truncated large distances
                     if (diff < thickness) {
                         // Brightness peaks at ring centre, falls off with distance from ring.
                         uint8_t falloff = static_cast<uint8_t>(((thickness - diff) * 255) / thickness);
